Middleware/respondio-middleware/tests/requirements.txt
# Mock MCP Server Requirements

fastapi>=0.109.0
uvicorn>=0.27.0
requests>=2.31.0
google-generativeai>=0.3.0
pydantic>=2.5.0
requests>=2.31.0
